+++ src/controllers/groupControllers/helpers/index.ts
@@ -1,0 +1,2 @@
+export * from './schemas';
+export * from './validates';
