+++ src/controllers/groupControllers/index.ts
@@ -1,2 +1,2 @@
 export { getListGroupsController } from './getListGroupsController';
-export { createNewGroupController } from './groupController';
+export { createNewGroupController } from './createNewGroupController';
