+++ src/controllers/messageControllers/index.ts
@@ -1,0 +1,1 @@
+export { getMessagesController } from './getMessagesController';
