+++ src/controllers/socketControllers/index.ts
@@ -1,0 +1,1 @@
+export { sendMessageController } from './sendMessageController';
