+++ src/routers/messageRouters.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 
-import { getMessagesController } from '@Controllers/messageControllers/getMessagesController';
+import { getMessagesController } from '@Controllers/messageControllers';
 import { verifyTokenMiddlewares } from '@Middlewares';
 
 const messageRouter = express.Router();
