+++ src/routers/userRouters.ts
@@ -4,7 +4,7 @@ import passport from 'passport';
 import { verifyTokenMiddlewares } from '@Middlewares';
 import { RequestUserAfterAuthenticate } from '@Models';
 
-import { getBaseUserController } from '../controllers/userControllers/getBaseUserController';
+import { getBaseUserController } from '../controllers/userControllers';
 import '../controllers/userControllers/userController';
 
 const userRouter = express.Router();
